+++ kwimage/structs/coords.py
@@ -293,9 +293,10 @@ class Coords(_generic.Spatial, ub.NiceRepr):
         Generalized coordinate transform.
 
         Args:
-            transform (GeometricTransform | ArrayLike | Augmenter):
-                scikit-image tranform, a transformation matrix, or
-                an imgaug Augmenter.
+            transform (GeometricTransform | ArrayLike | Augmenter | callable):
+                scikit-image tranform, a transformation matrix,
+                an imgaug Augmenter, or callable that maps the ndarray of
+                coords.
 
             input_dims (Tuple): shape of the image these objects correspond to
                 (only needed / used when transform is an imgaug augmenter)
@@ -325,15 +326,32 @@ class Coords(_generic.Spatial, ub.NiceRepr):
             >>> assert np.all(self.warp(np.eye(3)).data == self.data)
             >>> assert np.all(self.warp(np.eye(2)).data == self.data)
 
-        Ignore:
-            >>> # xdoctest: +SKIP
+        Doctest:
             >>> # xdoctest: +REQUIRES(module:osr)
+            >>> import osr
             >>> wgs84_crs = osr.SpatialReference()
             >>> wgs84_crs.ImportFromEPSG(4326)
-            >>> transform = osr.CoordinateTransformation(wgs84_crs, wgs84_crs)
+            >>> dst_crs = osr.SpatialReference()
+            >>> dst_crs.ImportFromEPSG(2927)
+            >>> transform = osr.CoordinateTransformation(wgs84_crs, dst_crs)
             >>> self = Coords.random(10, rng=0)
             >>> new = self.warp(transform)
-            >>> assert np.all(new.data == self.data)
+            >>> assert np.all(new.data != self.data)
+
+            >>> # Alternative using generic func
+            >>> def _gdal_coord_tranform(pts):
+            ...     return np.array([transform.TransformPoint(x, y, 0)[0:2]
+            ...                      for x, y in pts])
+            >>> alt = self.warp(_gdal_coord_tranform)
+            >>> assert np.all(alt.data != self.data)
+            >>> assert np.all(alt.data == new.data)
+
+        Doctest:
+            >>> # can use a generic function
+            >>> def func(xy):
+            ...     return np.zeros_like(xy)
+            >>> self = Points.random(10, rng=0)
+            >>> assert np.all(self.warp(func).xy == 0)
         """
         import kwimage
         impl = self._impl
@@ -369,24 +387,13 @@ class Coords(_generic.Spatial, ub.NiceRepr):
                         new_pts.append((x, y))
                     new.data = np.array(new_pts, dtype=new.data.dtype)
                     return new
+                elif callable(transform):
+                    new.data = transform(new.data)
+                    return new
             raise TypeError(type(transform))
         new.data = kwimage.warp_points(matrix, new.data)
         return new
 
-    def _warp_func(self, func, inplace=False):
-        """
-        Warp using a function that transforms points
-
-        Args:
-            func (callable): maps a single coordinate to a new coordinate
-        """
-        impl = self._impl
-        new = self if inplace else self.__class__(impl.copy(self.data), self.meta)
-        dtype = self.data.dtype
-        new.data = np.array([func(pt) for pt in new.data], dtype=dtype)
-        return new
-
-    # @profile
     def _warp_imgaug(self, augmenter, input_dims, inplace=False):
         """
         Warps by applying an augmenter from the imgaug library