+++ CHANGELOG.md
@@ -10,8 +10,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * `draw_line_segments_on_image`
 * Boxes.scale now accepts `about` keyword arg (can use to scale about center).
 * Boxes.warp now accepts matrices and does inexact corner warping
-* `kwimage.warp_points` now accepts `homog_mode`
-* Add `_warp_func`, which warps points using a general function to structures
+* kwimage structures `warp` function now accepts a generic callable for mapping array-based points.
 
 
 ### Changed